requirements.txt
control
cvxpy>=1.4.0
matplotlib
numpy
pandas
PyYAML
scipy